fix: added AdminCheck middleware to update and destroy route

+++ backend/routes/api.php
@@ -22,5 +22,5 @@ Route::group([
 Route::get('/cars', [CarController::class, 'index']);
 Route::get('/cars/{id}', [CarController::class, 'show']);
 Route::post('/cars', [CarController::class, 'store'])->middleware(AdminCheck::class);
-Route::put('cars/{id}', [CarController::class, 'update']);
-Route::delete('cars/{id}', [CarController::class, 'destroy']);
+Route::put('cars/{id}', [CarController::class, 'update'])->middleware(AdminCheck::class);
+Route::delete('cars/{id}', [CarController::class, 'destroy'])->middleware(AdminCheck::class);
